+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+# 1.2.0
+
+- Add `WidgetOnce` that return `Widget`
+- Add `runOnNewDay` method that runs on every new day.
+  The day here means you run the function at 3:00 AM. So, Day means
+  the next 12:00 AM.
+
 # 1.1.2
 
-- Down
+- Analyzing
 # 1.1.0
 
 - Fix bugs